terminar usuario y agregar validaciones

+++ src/user/user.controller.js
@@ -1,4 +1,5 @@
 import User from './user.model.js'
+import BuyCar from '../buyCar/buycar.model.js'
 import { encrypt, checkPassword, checkUserUpdate } from '../utils/validator.js'
 import { generateJwt } from '../utils/jwt.js'
 
@@ -21,7 +22,9 @@ export const saveClient = async (req, res) => {
         data.role = 'CLIENT'
         let user = new User(data)
         await user.save()
-        return res.send({ message: `user is saved seccesfully` })
+        let buyCar = new BuyCar({ user: user._id });
+        await buyCar.save();
+        return res.send({ message: `user is saved seccesfully and buy car for ${data.username} is added` })
     } catch (err) {
         console.error(err)
         return res.status(500).send({ message: 'Error saving user', err: err })
@@ -89,33 +92,78 @@ export const login = async (req, res) => {
 
 export const update = async (req, res) => {
     try {
+        const tokenId = req.user;
         let { id } = req.params
-        let data = req.body
-        let update = checkUserUpdate(data, id)
-        if (!update) return res.status(400).send({ message: 'Have submitted some data that cannot be updated or missing data' })
-        let updatedUser = await User.findOneAndUpdate(
-            { _id: id },
-            data,
-            { new: true }
-        )
-        //Validar la actualización
-        if (!updatedUser) return res.status(401).send({ message: 'User not found and not updated' })
-        //Respondo al usuario
-        return res.send({ message: 'Updated user', updatedUser })
+        if (tokenId._id.toString() === id || req.user.role === 'ADMIN') {
+            let data = req.body
+            let update = checkUserUpdate(data, id)
+            if (!update) return res.status(400).send({ message: 'Have submitted some data that cannot be updated or missing data' })
+            let updatedUser = await User.findOneAndUpdate(
+                { _id: id },
+                data,
+                { new: true }
+            )
+            //Validar la actualización
+            if (!updatedUser) return res.status(401).send({ message: 'User not found and not updated' })
+            //Respondo al usuario
+            return res.send({ message: 'Updated user', updatedUser })
+        }
+        return res.status(403).send({ message: 'Forbidden - You are not allowed or admin to update this account' });
     } catch (err) {
         console.error(err)
         return res.status(500).send({ message: 'Error updating account', err: err })
     }
 }
 
+export const updatePassword = async (req, res) => {
+    try {
+        const token = req.user;
+        const { id } = req.params;
+        if (token._id.toString() === id) {
+            const { newPassword, actualPassword } = req.body;
+
+            if (!actualPassword || !newPassword) {
+                return res.status(400).send({ message: 'Missing current or new password in request body' });
+            }
+            const originalPassword = checkPassword(token.password,
+                 actualPassword)
+            if (originalPassword) {
+                
+                const encriptedPassword = await encrypt(newPassword)
+                const updatedPassword = await User.findOneAndUpdate(
+                    { _id: id },
+                    { password: encriptedPassword },
+                    { new: true }
+                );
+                // Validar la actualización
+                if (!updatedPassword) {
+                    return res.status(401).send({ message: 'User not found or password not updated' });
+                }
+
+                // Responder al usuario
+                return res.send({ message: 'Updated password', updatedPassword })
+            }
+            return res.send({ message: 'this is not your actual password' })
+        }
+        return res.send({ message: 'this not is your account' })
+    } catch (err) {
+        console.error(err)
+        return res.status(500).send({ message: 'Error updating password', err: err })
+    }
+
+}
+
 export const erase = async (req, res) => {
     try {
         const tokenId = req.user;
         let { id } = req.params
+        let { validate } = req.body
+        if (!validate) return res.status(400).send({message: 'Write autorization word "CONFIRM"'})
+        if (validate !== 'CONFIRM') return res.status(400).send({message: 'This is not the authorization word "CONFIRM"'})
         if (tokenId._id.toString() === id || req.user.role === 'ADMIN') {
             let deletedUser = await User.findOneAndDelete({ _id: id })
-        if (!deletedUser) return res.status(404).send({ message: 'Account not found and not deleted' })
-        return res.send({ message: `Account with username ${deletedUser.username} deleted successfully` })
+            if (!deletedUser) return res.status(404).send({ message: 'Account not found and not deleted' })
+            return res.send({ message: `Account with username ${deletedUser.username} deleted successfully` })
         }
         return res.status(403).send({ message: 'Forbidden - You are not allowed to delete this account' });
